commented urls from css

+++ basic_app/ml_module.py
@@ -178,9 +178,9 @@ def run_job(username, job_code, domain, user_email):
     message.send()
 
     #clear ml waste(frames, video..)
-
-    job_frames_location = BASE_DIR / "framesimages/{}/{}/".format(username,job_code)
-    shutil.rmtree(job_frames_location)
+    if os.path.exists('framesimages/{}/{}'.format(username,job_code)):
+        job_frames_location = BASE_DIR / "framesimages/{}/{}/".format(username,job_code)
+        shutil.rmtree(job_frames_location)
     video_obj = Video.objects.filter(job_code=job_code).first()
     video_obj.delete()
